Integral versao 1 c/ semaforo

+++ adaquad.h
@@ -4,34 +4,34 @@
 //Structs
 typedef struct _static_worker_thread_arg {
 
-	double left_limit;
-	double right_limit;
-	double tolerance;
-	double (*calc_function) (double);
+	long double left_limit;
+	long double right_limit;
+	long double tolerance;
+	long double (*calc_function) (long double);
 
 } static_worker_thread_arg;
 
 //Functions to be the target of the integration operation
-double linear_function(double x);
+long double linear_function(long double x);
 
-double quadratic_function(double x);
+long double quadratic_function(long double x);
 
-double cubic_function(double x);
+long double cubic_function(long double x);
 
 //Function to calculate the area of a trapezoid
-double calc_trapezoid_area(double base1, double base2, double height);
+long double calc_trapezoid_area(long double base1, long double base2, long double height);
 
 //Functions to parallelize the operations
 
 //First Variant: each thread computes a subinterval for which will be responsible and calculates the result for this entire subinterval. When all threads have finished, the main thread should show the end result.
 
 //ptheads implementation
-void aq_static_administrator_sem_pthread(int num_threads, double left_limit, double right_limit, double tolerance, double (*calc_function) (double));
+void aq_static_administrator_sem_pthread(int num_threads, long double left_limit, long double right_limit, long double tolerance, long double (*calc_function) (long double));
 
 void* aq_static_worker_sem_pthread(void *arguments);
 
 //openmp implemantation
-void aq_static_administrator_sem_openmp(int num_threads, double left_limit, double right_limit, double tolerance, double (*calc_function) (double));
+void aq_static_administrator_sem_openmp(int num_threads, long double left_limit, long double right_limit, long double tolerance, long double (*calc_function) (long double));
 
 void* aq_static_worker_sem_openmp(void *arguments);
