+++ aqqueue.h
@@ -6,11 +6,11 @@
 //Elements of the queue
 typedef struct _interval interval ;
 struct _interval {
-	double left_limit;
-	double right_limit;
-	double f_left_limit;
-	double f_right_limit;
-	double area;
+	long double left_limit;
+	long double right_limit;
+	long double f_left_limit;
+	long double f_right_limit;
+	long double area;
 	interval *next;
 };
 
@@ -27,6 +27,6 @@ interval* dequeue(queue* q);
 
 queue* queue_initialize();
 
-interval* interval_initialize(double left_limit, double right_limit, double f_left_limit, double f_right_limit, double area);
+interval* interval_initialize(long double left_limit, long double right_limit, long double f_left_limit, long double f_right_limit, long double area);
 
 #endif
